Fix update function error behaviour, update demo data
