+++ src/JsonEditor.tsx
@@ -90,8 +90,10 @@ const Editor: React.FC<JsonEditorProps> = ({
       name: path.slice(-1)[0],
       path,
     })
-    if (result === false) return translate('ERROR_UPDATE', nodeData)
-    return result // Error string
+    if (result !== undefined) {
+      setData(currentData)
+      return result === false ? translate('ERROR_UPDATE', nodeData) : result
+    }
   }
 
   const onDelete: OnChangeFunction = async (value, path) => {
@@ -111,8 +113,10 @@ const Editor: React.FC<JsonEditorProps> = ({
       name: path.slice(-1)[0],
       path,
     })
-    if (result === false) return translate('ERROR_DELETE', nodeData)
-    return result // Error string
+    if (result !== undefined) {
+      setData(currentData)
+      return result === false ? translate('ERROR_UPDATE', nodeData) : result
+    }
   }
 
   const onAdd: OnChangeFunction = async (value, path) => {
@@ -132,8 +136,10 @@ const Editor: React.FC<JsonEditorProps> = ({
       name: path.slice(-1)[0],
       path,
     })
-    if (result === false) return translate('ERROR_ADD', nodeData)
-    return result // Error string
+    if (result !== undefined) {
+      setData(currentData)
+      return result === false ? translate('ERROR_UPDATE', nodeData) : result
+    }
   }
 
   const restrictEditFilter = getFilterFunction(restrictEdit)
